chaosface/pyproject.toml
[tool.poetry]
name = "chaosface"
version = "2.0.0"
description = "Chatbot and user interface for Twitch Controls Chaos"
authors = ["polysyl <polysyltwitch@gmail.com>", "blegas78"]
license = "GPL 3.0"
readme = "README.rst"

[tool.poetry.dependencies]
python = "^3.8"
twitchAPI = "*"
pyzmq = "*"
numpy = "*"
requests = "*"
fastapi = "*"
nicegui = "*"

[tool.poetry.dev-dependencies]
pytest = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
